moved prescan processing functions to l1_to_l2a.py
