+++ corgidrp/l1_to_l2a.py
@@ -1,21 +1,176 @@
 # A file that holds the functions that transmogrify l1 data to l2a data 
 
+import numpy as np
 
-def prescan_process(input_dataset): 
+image_constants= {
+    'SCI' : {
+        'frame_rows' : 1200,
+        'frame_cols' : 2200,
+        'image' : {
+            'rows': 1024,
+            'cols': 1024,
+            'r0c0': [13, 1088]
+            },
+        'prescan' : {
+            'rows': 1200,
+            'cols': 1088,
+            'r0c0': [0, 0]
+            },
+        'prescan_reliable' : {
+            'rows': 1200,
+            'cols': 200,
+            'r0c0': [0, 800]
+            },
+        'parallel_overscan' : {
+            'rows': 163,
+            'cols': 1056,
+            'r0c0': [1037, 1088]
+            },
+        'serial_overscan' : {
+            'rows': 1200,
+            'cols': 56,
+            'r0c0': [0, 2144]
+            },
+        },
+    'ENG' :{
+        'frame_rows' : 2200,
+        'frame_cols' : 2200,
+        'image' : {
+            'rows': 1024,
+            'cols': 1024,
+            'r0c0': [13, 1088]
+            },
+        'prescan' : {
+            'rows': 2200,
+            'cols': 1088,
+            'r0c0': [0, 0]
+            },
+        'prescan_reliable' : {
+            'rows': 2200,
+            'cols': 200,
+            'r0c0': [0, 800]
+            },
+        'parallel_overscan' : {
+            'rows': 1163,
+            'cols': 1056,
+            'r0c0': [1037, 1088]
+            },
+        'serial_overscan' : {
+            'rows': 2200,
+            'cols': 56,
+            'r0c0': [0, 2144]
+            },
+        },
+    }
+
+def slice_section(frame, obstype, key):
+    """Slice 2d section out of frame.
+
+    Parameters
+    ----------
+    frame : array_like
+        Full frame consistent with size given in frame_rows, frame_cols.
+    key : str
+        Keyword referencing section to be sliced; must exist in geom.
+
     """
-    
-    Crops the data arrays if appropriate and subtracts the prescan-bias
+    rows = image_constants[obstype][key]['rows']
+    cols = image_constants[obstype][key]['cols']
+    r0c0 = image_constants[obstype][key]['r0c0']
+
+    section = frame[r0c0[0]:r0c0[0]+rows, r0c0[1]:r0c0[1]+cols]
+    if section.size == 0:
+        raise Exception('Corners invalid')
+    return section
+
+def prescan_biassub(input_dataset, bias_offset=0., return_full_frame=False):
+    """
+    Perform pre-scan bias subtraction of a dataset.
 
     Args:
-        input_dataset (corgidrp.data.Dataset): a dataset of Images that need prescan processing (L1-level)
-
+        input_dataset (corgidrp.data.Dataset): a dataset of Images (L1a-level)
+        bias_offset (float): an offset value to be subtracted from the bias
+        return_full_frame (bool): flag indicating whether to return the full frame or only the bias-subtracted image area
     Returns:
-        corgidrp.data.Dataset: a prescan-processed version of the input dataset
+        corgidrp.data.Dataset: a pre-scan bias subtracted version of the input dataset
     """
+    # Make a copy of the input dataset to operate on
+    output_dataset = input_dataset.copy()
 
-    return None
+    #initialize list of output frames to be concatenated
+    out_frames_data = []
+    out_frames_err = []
+    out_frames_dq = []
 
+    # Iterate over frames
+    for frame in output_dataset:
 
+        frame_data = frame.data
+        frame_err = frame.err
+        frame_dq = frame.dq
+
+        # Determine what type of file it is (engineering or science), then choose detector area dict
+        obstype = frame.pri_hdr['OBSTYPE']
+        assert obstype in ['SCI','ENG'], f"Observation type of frame {i} is not 'SCI' or 'ENG'"
+
+        # Get the reliable prescan area
+        prescan = slice_section(frame_data, obstype, 'prescan_reliable')
+
+        if not return_full_frame:
+            # Get the image area
+            image_data = slice_section(frame_data, obstype, 'image')
+            image_err = slice_section(frame_err, obstype, 'image')
+            image_dq = slice_section(frame_dq, obstype, 'image')
+            
+            # Get the part of the prescan that lines up with the image, and do a
+            # row-by-row bias subtraction on it
+            i_r0 = image_constants[obstype]['image']['r0c0'][0]
+            p_r0 = image_constants[obstype]['prescan']['r0c0'][0]
+            i_nrow = image_constants[obstype]['image']['rows']
+     
+            # Get data from prescan (alined with image area)
+            al_prescan = prescan[(i_r0-p_r0):(i_r0-p_r0+i_nrow), :]    
+            medbyrow = np.median(al_prescan, axis=1)[:, np.newaxis]
+        else:
+            image_data = frame_data
+            image_err = frame_err
+            image_dq = frame_dq
+            
+            medbyrow = np.median(prescan, axis=1)[:, np.newaxis]  
+
+        bias = medbyrow - bias_offset
+        image_bias_corrected = image_data - bias
+
+        out_frames_data.append(image_bias_corrected)
+        out_frames_err.append(image_err)
+        out_frames_dq.append(image_dq)
+
+        # Update header in the dataset
+        frame.ext_hdr['NAXIS1'] = image_bias_corrected.shape[1]
+        frame.ext_hdr['NAXIS2'] = image_bias_corrected.shape[0]
+
+    
+    # Update all_data and reassign frame pointers (only necessary because the array size has changed)
+    out_frames_data_arr = np.array(out_frames_data)
+    out_frames_err_arr = np.array(out_frames_err)
+    out_frames_dq_arr = np.array(out_frames_dq)
+
+    output_dataset.all_data = out_frames_data_arr
+    output_dataset.all_err = out_frames_err_arr
+    output_dataset.all_dq = out_frames_dq_arr
+
+    for i,frame in enumerate(output_dataset):
+        frame.data = out_frames_data_arr[i]
+        frame.err = out_frames_err_arr[i]
+        frame.dq = out_frames_dq_arr[i]
+
+            
+    history_msg = "Frames cropped and bias subtracted"
+
+    # update the output dataset with this new dark subtracted data and update the history
+    output_dataset.update_after_processing_step(history_msg)
+
+    return output_dataset
 
 def detect_cosmic_rays(input_dataset):
     """
@@ -44,18 +199,4 @@ def correct_nonlinearity(input_dataset, non_lin_correction):
         corgidrp.data.Dataset: a non-linearity corrected version of the input dataset
     """
 
-    return None
-
-def prescan_biassub(input_dataset, bias_offset=0., return_full_frame=False):
-    """
-    Perform pre-scan bias subtraction of a dataset.
-
-    Args:
-        input_dataset (corgidrp.data.Dataset): a dataset of Images (L1a-level)
-        bias_offset (float): an offset value to be subtracted from the bias
-        return_full_frame (bool, optional): flag indicating whether to return the full frame or only the bias-subtracted image area
-    
-    Returns:
-        corgidrp.data.Dataset: a pre-scan bias subtracted version of the input dataset
-    """
     return None
